w3c-reports/12092014/efl/runned-tests/cors/credentials-flag-expected.txt

CONSOLE MESSAGE: line 80: NO_COOKIE_
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=True. Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=TRUE. Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=%22true%22. Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=0. Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=false. Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=1. Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CORS - Access-Control-Allow-Credentials


FAIL Setting withCredentials on a sync XHR object should throw assert_throws: setting withCredentials function "function () { client.withCredentials = true; }" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/cors/credentials-flag.htm:23:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/cors/credentials-flag.htm:20:5)
PASS Don't send cookie by default 
FAIL Don't send cookie part 2 assert_equals: Cookie sent in withCredentials=true sync request expected "COOKIE" but got "NO_COOKIE"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/credentials-flag.htm:62:30
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
FAIL Don't obey Set-Cookie when withCredentials=false assert_equals: third expected "COOKIE" but got "NO_COOKIE"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/credentials-flag.htm:93:30
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS Access-Control-Allow-Credentials: TRUE should be disallowed (async) 
PASS Access-Control-Allow-Credentials: True should be disallowed (async) 
PASS Access-Control-Allow-Credentials: "true" should be disallowed (async) 
PASS Access-Control-Allow-Credentials: false should be disallowed (async) 
PASS Access-Control-Allow-Credentials: 1 should be disallowed (async) 
PASS Access-Control-Allow-Credentials: 0 should be disallowed (async) 
